backend-python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow>=10.0.0
pyttsx3==2.90
gtts==2.5.3
requests==2.32.3
